Call CLI from main
- Extract `start-env` into its own function, which can be called via
  `StructOp`.
