+++ src/create_comit_app.rs
@@ -3,11 +3,5 @@ use structopt::StructOpt;
 #[derive(StructOpt, Debug)]
 #[structopt(name = "create-comit-app")]
 pub enum CreateComitApp {
-    StartEnv {},
+    StartEnv,
 }
-
-// pub fn start_btsieve() {
-//     Command::new("btsieve")
-//         .spawn()
-//         .expect("btsieve not found in path");
-// }
