+++ src/lib.rs
@@ -3,3 +3,4 @@ pub mod btsieve;
 pub mod cnd;
 pub mod create_comit_app;
 pub mod ethereum;
+pub mod start_env;
